refactor: consolidate error handling into a single error page and update routing
